requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
networkx>=3.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
tiktoken>=0.5.0
scikit-learn>=1.3.0